Refactor SubAccount class and tests to move from Organization namespace to General namespace
