--- src/Api/Organization/OrganizationInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Mailtrap\Api\Organization;
-
-interface OrganizationInterface
-{
-}
